Restructure project as package
