+++ tests/test_client_update_dns.py
@@ -1,11 +1,7 @@
-import os, sys
-
-sys.path.insert(
-    0, os.path.abspath(os.path.join(os.path.dirname(__file__), ".."))
-)
-from client import update_dns
+from hetzner_dyndns.client import update_dns
 import pytest
 import requests
+import sys
 
 
 def test_get_verify_option_ca_bundle(monkeypatch):
